requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic-settings==2.5.2
SQLAlchemy==2.0.36
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
greenlet==3.1.1

# Testing
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
